# api/documentation/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This project is a Node.js api to create your own IMDB. Users are able to add movies, actors and reviews.",
    "version": "1.0.0",
    "title": "Private IMDB API"
  },
...
  "paths": {
...
    "/professionals/create-professional": {
      "put": {
        "tags": [
          "professionals"
        ],
        "summary": "Create a professional",
        "operationId": "createProfessional",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "professional data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "professional": {
                  "$ref": "#/definitions/professionals"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "insertId": {
                  "type": "integer"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, you need a token access to execute this request"
          },
          "412": {
            "description": "Invalid input, missing some parameters in path or body"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/professionals/update-professional": {
      "put": {
        "tags": [
          "professionals"
        ],
        "summary": "Update a professional",
        "operationId": "updateProfessional",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "professional data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "professional": {
                  "$ref": "#/definitions/professionals"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "401": {
            "description": "Unauthorized, you need a token access to execute this request"
          },
          "412": {
            "description": "Invalid input, missing some parameters in path or body"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/professionals/name/{name}": {
      "get": {
        "tags": [
          "professionals"
        ],
        "summary": "Find a professional by name",
        "operationId": "getProfessionalByName",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "description": "the name of the professional (firstname + lastname)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/professionals"
            }
          },
          "401": {
            "description": "Unauthorized, you need a token access to execute this request"
          },
          "412": {
            "description": "Invalid input, missing some parameters in path or body"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/professionals/{id}": {
      "delete": {
        "tags": [
          "professionals"
        ],
        "summary": "Delete a professional by id",
        "operationId": "deleteProfessionalById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "the id of the professional to delete",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "401": {
            "description": "Unauthorized, you need a token access to execute this request"
          },
          "412": {
            "description": "Invalid input, missing some parameters in path or body"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "professionals"
        ],
        "summary": "Find a professional by id",
        "operationId": "getProfessionalById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "the id of the professional to find",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/professionals"
            }
          },
          "401": {
            "description": "Unauthorized, you need a token access to execute this request"
          },
          "412": {
            "description": "Invalid input, missing some parameters in path or body"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/professionals/with-details/{id}": {
      "get": {
        "tags": [
          "professionals"
        ],
        "summary": "Find a professional by id with all associations",
        "operationId": "getProfessionalByIdWithAssociations",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "the id of the professional to find",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "firstname": {
                  "type": "string"
                },
                "lastname": {
                  "type": "string"
                },
                "profilePicture": {
                  "type": "string"
                },
                "birthDate": {
                  "type": "string"
                },
                "nationality": {
                  "type": "string"
                },
                "job": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "string"
                },
                "updatedAt": {
                  "type": "string"
                },
                "movies": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/movies"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, you need a token access to execute this request"
          },
          "412": {
            "description": "Invalid input, missing some parameters in path or body"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/reviews/create-review": {
      "put": {
        "tags": [
          "reviews"
        ],
        "summary": "Create a review",
        "operationId": "createReview",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "review data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "review": {
                  "$ref": "#/definitions/reviews"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "insertId": {
                  "type": "integer"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, you need a token access to execute this request"
          },
          "412": {
            "description": "Invalid input, missing some parameters in path or body"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/reviews/update-review": {
      "put": {
        "tags": [
          "reviews"
        ],
        "summary": "Update a review, only the review author can do this",
        "operationId": "updateReview",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "review data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "review": {
                  "$ref": "#/definitions/reviews"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "401": {
            "description": "Unauthorized, you need a token access to execute this request"
          },
          "412": {
            "description": "Invalid input, missing some parameters in path or body"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/reviews/{id}": {
      "delete": {
        "tags": [
          "reviews"
        ],
        "summary": "Delete a review by id, only the review author can do this",
        "operationId": "deleteReviewById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "the id of the review to delete",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "401": {
            "description": "Unauthorized, you need a token access to execute this request"
          },
          "412": {
            "description": "Invalid input, missing some parameters in path or body"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "Find a review by id",
        "operationId": "getReviewById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "the id of the review to find",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/reviews"
            }
          },
          "401": {
            "description": "Unauthorized, you need a token access to execute this request"
          },
          "412": {
            "description": "Invalid input, missing some parameters in path or body"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/reviews/with-details/{id}": {
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "Find a review by id with all associations",
        "operationId": "getReviewByIdWithAssociations",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "the id of the review to find",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "rate": {
                  "type": "number"
                },
                "review": {
                  "type": "string"
                },
                "userId": {
                  "type": "integer"
                },
                "movieId": {
                  "type": "integer"
                },
                "createdAt": {
                  "type": "string"
                },
                "updatedAt": {
                  "type": "string"
                },
                "movie": {
                  "$ref": "#/definitions/movies"
                },
                "user": {
                  "$ref": "#/definitions/users"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, you need a token access to execute this request"
          },
          "412": {
            "description": "Invalid input, missing some parameters in path or body"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "movies": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "rate": {
          "type": "number"
        },
        "date": {
          "type": "string"
        },
        "duration": {
          "type": "integer"
        },
        "genre": {
          "type": "string"
        },
        "synopsis": {
          "type": "string"
        },
        "imageUrl": {
          "type": "string"
        },
        "teaserUrl": {
          "type": "string"
        },
        "nationality": {
          "type": "string"
        },
        "userId": {
          "type": "integer"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
    "professionals": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "profilePicture": {
          "type": "string"
        },
        "birthDate": {
          "type": "string"
        },
        "nationality": {
          "type": "string"
        },
        "job": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
    "reviews": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "rate": {
          "type": "number"
        },
        "review": {
          "type": "string"
        },
        "userId": {
          "type": "integer"
        },
        "movieId": {
          "type": "integer"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
    "users": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "profilePicture": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    }
  }
}